COL-1783, delete existing fabric elements via socket.io emit
